Prevent arguments from being interpreted as options (#80)
Prevent errors that can happen if the current token, selected path, or selected command, beings with a `-`. 
Examples where fzf.fish would error:
- the cursor is over the token --something when the user executes search files
- the cursor is over -z when the user executes search variables
- the selected changed path from git status is -folder
- the selected command is --function

The solution is to delineate the end of options with --, which is a POSIX specification defined in https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap12.html#tag_12_02 (guideline 10), which is respected by Fish builtins. After --, the command will always interpret the arguments as positional arguments.