Fix search file feature unnecessarily prepending ./ (#119)
If the file path already starts with . or /, which can happen if such a path was used as the base directory, then don't prepend ./. Without such logic, the path becomes invalid. For example:
- if the current token is /etc/ and you select systemd, you end up with './/etc/systemd'
- if the current token is ~/ and you select Desktop, you end up with './~/desktop'

Additionally, since the output paths are quoted, we need to use the expanded token when outputting paths because ~ doesn't get expanded inside quotes.